feat: 스킬 클래스 분리
